pyproject.toml
[project]
name = "lm-as-sim-coding-tasks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.52.2",
    "openai>=1.84.0",
    "pygithub>=2.6.1",
    "python-dotenv>=1.1.0",
    "together>=1.5.8",
]
